Make hydrateConfigurationStore more reusable
