+++ src/configuration-requestor.ts
@@ -1,14 +1,7 @@
 import { IConfigurationStore } from './configuration-store/configuration-store';
+import { hydrateConfigurationStore } from './configuration-store/configuration-store-utils';
 import { IHttpClient } from './http-client';
-import {
-  BanditVariation,
-  BanditParameters,
-  Flag,
-  Environment,
-  PrecomputedFlag,
-} from './interfaces';
-
-type Entry = Flag | BanditVariation[] | BanditParameters | PrecomputedFlag;
+import { BanditVariation, BanditParameters, Flag } from './interfaces';
 
 // Requests AND stores flag configurations
 export default class ConfigurationRequestor {
@@ -18,7 +11,6 @@ export default class ConfigurationRequestor {
     private readonly banditVariationConfigurationStore: IConfigurationStore<
       BanditVariation[]
     > | null,
-    private readonly precomputedFlagStore: IConfigurationStore<PrecomputedFlag>,
     private readonly banditModelConfigurationStore: IConfigurationStore<BanditParameters> | null,
   ) {}
 
@@ -28,7 +20,7 @@ export default class ConfigurationRequestor {
       return;
     }
 
-    await this.hydrateConfigurationStore(this.flagConfigurationStore, {
+    await hydrateConfigurationStore(this.flagConfigurationStore, {
       entries: configResponse.flags,
       environment: configResponse.environment,
       createdAt: configResponse.createdAt,
@@ -42,7 +34,7 @@ export default class ConfigurationRequestor {
       // Map bandit flag associations by flag key for quick lookup (instead of bandit key as provided by the UFC)
       const banditVariations = this.indexBanditVariationsByFlagKey(configResponse.bandits);
 
-      await this.hydrateConfigurationStore(this.banditVariationConfigurationStore, {
+      await hydrateConfigurationStore(this.banditVariationConfigurationStore, {
         entries: banditVariations,
         environment: configResponse.environment,
         createdAt: configResponse.createdAt,
@@ -55,29 +47,11 @@ export default class ConfigurationRequestor {
           throw new Error('Bandit parameters fetched but no bandit configuration store provided');
         }
 
-        await this.hydrateConfigurationStore(this.banditModelConfigurationStore, {
+        await hydrateConfigurationStore(this.banditModelConfigurationStore, {
           entries: banditResponse.bandits,
           environment: configResponse.environment,
           createdAt: configResponse.createdAt,
         });
-      }
-    }
-  }
-
-  private async hydrateConfigurationStore<T extends Entry>(
-    configurationStore: IConfigurationStore<T> | null,
-    response: {
-      entries: Record<string, T>;
-      environment: Environment;
-      createdAt: string;
-    },
-  ): Promise<void> {
-    if (configurationStore) {
-      const didUpdate = await configurationStore.setEntries(response.entries);
-      if (didUpdate) {
-        configurationStore.setEnvironment(response.environment);
-        configurationStore.setConfigFetchedAt(new Date().toISOString());
-        configurationStore.setConfigPublishedAt(response.createdAt);
       }
     }
   }
@@ -97,18 +71,5 @@ export default class ConfigurationRequestor {
       });
     });
     return banditVariationsByFlagKey;
-  }
-
-  async fetchAndStorePrecomputedFlags(): Promise<void> {
-    const precomputedResponse = await this.httpClient.getPrecomputedFlags();
-    if (!precomputedResponse?.flags) {
-      return;
-    }
-
-    await this.hydrateConfigurationStore(this.precomputedFlagStore, {
-      entries: precomputedResponse.flags,
-      environment: precomputedResponse.environment,
-      createdAt: precomputedResponse.createdAt,
-    });
   }
 }
