+++ src/configuration-store/configuration-store-utils.ts
@@ -1,0 +1,29 @@
+import {
+  BanditParameters,
+  BanditVariation,
+  Environment,
+  Flag,
+  PrecomputedFlag,
+} from '../interfaces';
+
+import { IConfigurationStore } from './configuration-store';
+
+type Entry = Flag | BanditVariation[] | BanditParameters | PrecomputedFlag;
+
+export async function hydrateConfigurationStore<T extends Entry>(
+  configurationStore: IConfigurationStore<T> | null,
+  response: {
+    entries: Record<string, T>;
+    environment: Environment;
+    createdAt: string;
+  },
+): Promise<void> {
+  if (configurationStore) {
+    const didUpdate = await configurationStore.setEntries(response.entries);
+    if (didUpdate) {
+      configurationStore.setEnvironment(response.environment);
+      configurationStore.setConfigFetchedAt(new Date().toISOString());
+      configurationStore.setConfigPublishedAt(response.createdAt);
+    }
+  }
+}
